Relocate database tests
